+++ features/admin.core.v1/components/pre-loader/pre-loader.tsx
@@ -16,12 +16,28 @@
  * under the License.
  */
 
+import { Theme, styled } from "@mui/material/styles";
+import Box from "@oxygen-ui/react/Box";
 import { commonConfig } from "@wso2is/admin.extensions.v1";
 import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import { ContentLoader } from "@wso2is/react-components";
 import React, { FunctionComponent, ReactElement } from "react";
 import { ReactComponent as Pulse } from "./pulse.svg";
-import "./pre-loader.css";
+
+const PulseLoader: typeof Box = styled(Box)(({ theme }: { theme: Theme }) => ({
+    "& svg": {
+        animation: "app-pre-loader-pulse 2s ease-in-out infinite",
+        height: theme.spacing(10),
+        width: theme.spacing(10)
+    },
+    "@keyframes app-pre-loader-pulse": {
+        "0%, 100%": { opacity: 1 },
+        "50%": { opacity: 0.5 }
+    },
+    alignItems: "center",
+    display: "flex",
+    justifyContent: "center"
+}));
 
 /**
  * Pre loader component props interface.
@@ -48,9 +64,9 @@ export const PreLoader: FunctionComponent<PreLoaderPropsInterface> = (
             { commonConfig?.enableDefaultPreLoader ? (
                 <ContentLoader data-componentid={ componentId } />
             ) : (
-                <div className="app-pre-loader" data-componentid={ componentId }>
+                <PulseLoader data-componentid={ componentId }>
                     <Pulse data-componentid={ `${componentId}-svg` } />
-                </div>
+                </PulseLoader>
             ) }
         </div>
     );